phase_12: Remastered Cog Paintings!
This took a lot of work to do, and it's still not ideal (some of the effects don't look right, for example) but it's about as perfect as you can get without the original art on hand.

The scans were taken from Footballtoon's since TTR's doesn't have many of the Cogs needed, and were then digitally modified to make them as close as I can get them to the originals.