phase_3.5: Upscaled wall_lrg_brick
This looks a tad inconsistent in the original palette due to the original being a mere 128x128. My guess is this was supposed to be 256x256, but Disney being Disney means it was downscaled pre-repository inclusion. Upscaled it back to its "correct" look, and it looks a lot cleaner as a result.